Re-org the processors into one module.
