--- pypond/mapper.py
@@ -1,68 +1,0 @@
-#  Copyright (c) 2016, The Regents of the University of California,
-#  through Lawrence Berkeley National Laboratory (subject to receipt
-#  of any required approvals from the U.S. Dept. of Energy).
-#  All rights reserved.
-#
-#  This source code is licensed under the BSD-style license found in the
-#  LICENSE file in the root directory of this source tree.
-
-"""
-Map operator
-"""
-
-from .pipeline_io import PipelineIOException
-from .sources import Processor
-from .util import is_pipeline, Options
-
-
-class Mapper(Processor):
-    """
-    A processor which takes an operator as its only option
-    and uses that to either output a new event.
-
-    Parameters
-    ----------
-    arg1 : Mapper or Pipeline
-        Copy constructor or the pipeline.
-    options : Options
-        Options object.
-    """
-
-    def __init__(self, arg1, options=Options()):
-        """create the mapper"""
-
-        super(Mapper, self).__init__(arg1, options)
-
-        self._log('Mapper.init', 'uid: {0}'.format(self._id))
-
-        self._op = None
-
-        if isinstance(arg1, Mapper):
-            self._op = arg1._op  # pylint: disable=protected-access
-        elif is_pipeline(arg1):
-            self._op = options.op
-        else:
-            msg = 'Unknown arg to Mapper: {0}'.format(arg1)
-            raise PipelineIOException(msg)
-
-        if callable(self._op) is False:
-            msg = 'op: {0} is not a callable function'.format(self._op)
-            raise PipelineIOException(msg)
-
-    def clone(self):
-        """clone it."""
-        return Mapper(self)
-
-    def add_event(self, event):
-        """
-        Perform the map operation on the event and emit.
-
-        Parameters
-        ----------
-        event : Event, IndexedEvent, TimerangeEvent
-            Any of the three event variants.
-        """
-        if self.has_observers():
-            evn = self._op(event)
-            self._log('Mapper.add_event', 'emitting: {0}'.format(evn))
-            self.emit(evn)
